+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/my/CreateChannelDialog.scala
@@ -8,10 +8,11 @@ import android.support.v7.app.{AlertDialog, AppCompatDialogFragment}
 import android.view.inputmethod.InputMethodManager
 import android.widget.Button
 import x7c1.linen.database.control.DatabaseHelper
-import x7c1.linen.database.struct.{ChannelParts, ChannelSubscriber}
+import x7c1.linen.database.struct.ChannelParts
 import x7c1.linen.glue.res.layout.SettingMyChannelCreate
 import x7c1.linen.modern.init.settings.my.CreateChannelDialog.Arguments
 import x7c1.linen.repository.account.AccountIdentifiable
+import x7c1.linen.repository.channel.subscribe.ChannelSubscriber
 import x7c1.linen.repository.date.Date
 import x7c1.wheat.ancient.context.ContextualFactory
 import x7c1.wheat.ancient.resource.ViewHolderProviderFactory
