+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/my/MyChannelsDelegatee.scala
@@ -6,13 +6,13 @@ import android.support.v4.app.FragmentActivity
 import android.support.v7.app.AlertDialog
 import android.support.v7.widget.LinearLayoutManager
 import x7c1.linen.database.control.DatabaseHelper
-import x7c1.linen.database.struct.ChannelSubscriber
 import x7c1.linen.glue.activity.ActivityControl
 import x7c1.linen.glue.activity.ActivityLabel.SettingMyChannelSources
 import x7c1.linen.glue.res.layout.{SettingMyChannelCreate, SettingMyChannelsLayout}
 import x7c1.linen.modern.display.settings.{ChannelRowAdapter, ChannelSourcesSelected, MyChannelSubscriptionChanged}
 import x7c1.linen.repository.account.{AccountIdentifiable, ClientAccount}
 import x7c1.linen.repository.channel.my.{MyChannelAccessor, MyChannelAccessorLoader}
+import x7c1.linen.repository.channel.subscribe.ChannelSubscriber
 import x7c1.wheat.ancient.context.ContextualFactory
 import x7c1.wheat.ancient.resource.ViewHolderProviderFactory
 import x7c1.wheat.lore.resource.AdapterDelegatee
