+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/PresetsAllFragment.scala
@@ -4,10 +4,10 @@ import android.content.Context
 import android.os.Bundle
 import android.view.{LayoutInflater, View, ViewGroup}
 import x7c1.linen.database.control.DatabaseHelper
-import x7c1.linen.database.struct.ChannelSubscriber
 import x7c1.linen.glue.res.layout.{SettingPresetChannelRow, SettingPresetTabAll}
 import x7c1.linen.repository.account.AccountIdentifiable
 import x7c1.linen.repository.channel.preset.AllPresetChannelsAccessor
+import x7c1.linen.repository.channel.subscribe.ChannelSubscriber
 import x7c1.wheat.ancient.resource.ViewHolderProviderFactory
 import x7c1.wheat.macros.fragment.TypedFragment
 import x7c1.wheat.macros.intent.LocalBroadcaster
