+++ linen-repository/src/test/scala/x7c1/linen/repository/channel/preset/PresetChannelsAccessorTest.scala
@@ -7,8 +7,8 @@ import org.robolectric.annotation.Config
 import org.robolectric.{RobolectricTestRunner, RuntimeEnvironment}
 import org.scalatest.junit.JUnitSuiteLike
 import x7c1.linen.database.control.DatabaseHelper
-import x7c1.linen.database.struct.ChannelSubscriber
 import x7c1.linen.repository.account.setup.ClientAccountSetup
+import x7c1.linen.repository.channel.subscribe.ChannelSubscriber
 import x7c1.linen.repository.preset.PresetFactory
 
 @Config(manifest=Config.NONE)
